+++ hydro_lang/src/stream.rs
@@ -1898,53 +1898,6 @@ where
             },
         )
     }
-
-    /// A special case of [`Stream::reduce_keyed`] where tuples with keys less than the watermark are automatically deleted.
-    ///
-    /// The input stream must have a [`TotalOrder`] guarantee, which means that the `comb` closure is allowed
-    /// to depend on the order of elements in the stream.
-    ///
-    /// # Example
-    /// ```rust
-    /// # use hydro_lang::*;
-    /// # use futures::StreamExt;
-    /// # tokio_test::block_on(test_util::stream_transform_test(|process| {
-    /// let tick = process.tick();
-    /// let watermark = tick.singleton(q!(1));
-    /// let numbers = process.source_iter(q!([(0, 100), (1, 101), (2, 102), (2, 102)]));
-    /// let batch = unsafe { numbers.tick_batch(&tick) };
-    /// batch
-    ///     .reduce_keyed_watermark(watermark, q!(|acc, x| *acc += x))
-    ///     .all_ticks()
-    /// # }, |mut stream| async move {
-    /// // (2, 204)
-    /// # assert_eq!(stream.next().await.unwrap(), (2, 204));
-    /// # }));
-    /// ```
-    pub fn reduce_keyed_watermark<O, F>(
-        self,
-        other: impl Into<Optional<O, Tick<L>, Bounded>>,
-        comb: impl IntoQuotedMut<'a, F, Tick<L>>,
-    ) -> Stream<(K, V), Tick<L>, Bounded, NoOrder, ExactlyOnce>
-    where
-        O: Clone,
-        F: Fn(&mut V, V) + 'a,
-        K: Default, // So we can have a default watermark
-    {
-        let other: Optional<O, Tick<L>, Bounded> = other.into();
-        check_matching_location(&self.location, &other.location);
-        let f = comb.splice_fn2_borrow_mut_ctx(&self.location).into();
-
-        Stream::new(
-            self.location.clone(),
-            HydroNode::ReduceKeyedWatermark {
-                f,
-                input: Box::new(self.ir_node.into_inner()),
-                watermark: Box::new(other.ir_node.into_inner()),
-                metadata: self.location.new_node_metadata::<(K, V)>(),
-            },
-        )
-    }
 }
 
 impl<'a, K, V, L, O, R> Stream<(K, V), Tick<L>, Bounded, O, R>
@@ -2054,45 +2007,6 @@ where
             .reduce_commutative_idempotent(comb)
             .entries()
     }
-
-    /// A special case of [`Stream::reduce_keyed_commutative_idempotent`] where tuples with keys less than the watermark are automatically deleted.
-    ///
-    /// The `comb` closure must be **commutative**, as the order of input items is not guaranteed, and **idempotent**,
-    /// as there may be non-deterministic duplicates.
-    ///
-    /// # Example
-    /// ```rust
-    /// # use hydro_lang::*;
-    /// # use futures::StreamExt;
-    /// # tokio_test::block_on(test_util::stream_transform_test(|process| {
-    /// let tick = process.tick();
-    /// let watermark = tick.singleton(q!(1));
-    /// let numbers = process.source_iter(q!([(0, false), (1, false), (2, false), (2, true)]));
-    /// let batch = unsafe { numbers.tick_batch(&tick) };
-    /// batch
-    ///     .reduce_keyed_watermark_commutative_idempotent(watermark, q!(|acc, x| *acc |= x))
-    ///     .all_ticks()
-    /// # }, |mut stream| async move {
-    /// // (2, true)
-    /// # assert_eq!(stream.next().await.unwrap(), (2, true));
-    /// # }));
-    /// ```
-    pub fn reduce_keyed_watermark_commutative_idempotent<O2, F>(
-        self,
-        other: impl Into<Optional<O2, Tick<L>, Bounded>>,
-        comb: impl IntoQuotedMut<'a, F, Tick<L>>,
-    ) -> Stream<(K, V), Tick<L>, Bounded, NoOrder, ExactlyOnce>
-    where
-        O2: Clone,
-        F: Fn(&mut V, V) + 'a,
-        K: Default,
-    {
-        unsafe {
-            // SAFETY: the combinator function is commutative and idempotent
-            self.assume_ordering().assume_retries()
-        }
-        .reduce_keyed_watermark(other, comb)
-    }
 }
 
 impl<'a, K, V, L, O> Stream<(K, V), Tick<L>, Bounded, O, ExactlyOnce>
@@ -2174,44 +2088,6 @@ where
     {
         self.into_keyed().reduce_commutative(comb).entries()
     }
-
-    /// A special case of [`Stream::reduce_keyed_commutative`] where tuples with keys less than the watermark are automatically deleted.
-    ///
-    /// The `comb` closure must be **commutative**, as the order of input items is not guaranteed.
-    ///
-    /// # Example
-    /// ```rust
-    /// # use hydro_lang::*;
-    /// # use futures::StreamExt;
-    /// # tokio_test::block_on(test_util::stream_transform_test(|process| {
-    /// let tick = process.tick();
-    /// let watermark = tick.singleton(q!(1));
-    /// let numbers = process.source_iter(q!([(0, 100), (1, 101), (2, 102), (2, 102)]));
-    /// let batch = unsafe { numbers.tick_batch(&tick) };
-    /// batch
-    ///     .reduce_keyed_watermark_commutative(watermark, q!(|acc, x| *acc += x))
-    ///     .all_ticks()
-    /// # }, |mut stream| async move {
-    /// // (2, 204)
-    /// # assert_eq!(stream.next().await.unwrap(), (2, 204));
-    /// # }));
-    /// ```
-    pub fn reduce_keyed_watermark_commutative<O2, F>(
-        self,
-        other: impl Into<Optional<O2, Tick<L>, Bounded>>,
-        comb: impl IntoQuotedMut<'a, F, Tick<L>>,
-    ) -> Stream<(K, V), Tick<L>, Bounded, NoOrder, ExactlyOnce>
-    where
-        O2: Clone,
-        F: Fn(&mut V, V) + 'a,
-        K: Default,
-    {
-        unsafe {
-            // SAFETY: the combinator function is commutative
-            self.assume_ordering()
-        }
-        .reduce_keyed_watermark(other, comb)
-    }
 }
 
 impl<'a, K, V, L, R> Stream<(K, V), Tick<L>, Bounded, TotalOrder, R>
@@ -2292,44 +2168,6 @@ where
         F: Fn(&mut V, V) + 'a,
     {
         self.into_keyed().reduce_idempotent(comb).entries()
-    }
-
-    /// A special case of [`Stream::reduce_keyed_idempotent`] where tuples with keys less than the watermark are automatically deleted.
-    ///
-    /// The `comb` closure must be **idempotent**, as there may be non-deterministic duplicates.
-    ///
-    /// # Example
-    /// ```rust
-    /// # use hydro_lang::*;
-    /// # use futures::StreamExt;
-    /// # tokio_test::block_on(test_util::stream_transform_test(|process| {
-    /// let tick = process.tick();
-    /// let watermark = tick.singleton(q!(1));
-    /// let numbers = process.source_iter(q!([(0, false), (1, false), (2, false), (2, true)]));
-    /// let batch = unsafe { numbers.tick_batch(&tick) };
-    /// batch
-    ///     .reduce_keyed_watermark_idempotent(watermark, q!(|acc, x| *acc |= x))
-    ///     .all_ticks()
-    /// # }, |mut stream| async move {
-    /// // (2, true)
-    /// # assert_eq!(stream.next().await.unwrap(), (2, true));
-    /// # }));
-    /// ```
-    pub fn reduce_keyed_watermark_idempotent<O2, F>(
-        self,
-        other: impl Into<Optional<O2, Tick<L>, Bounded>>,
-        comb: impl IntoQuotedMut<'a, F, Tick<L>>,
-    ) -> Stream<(K, V), Tick<L>, Bounded, NoOrder, ExactlyOnce>
-    where
-        O2: Clone,
-        F: Fn(&mut V, V) + 'a,
-        K: Default,
-    {
-        unsafe {
-            // SAFETY: the combinator function is idempotent
-            self.assume_retries()
-        }
-        .reduce_keyed_watermark(other, comb)
     }
 }
 
@@ -2636,7 +2474,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use futures::{SinkExt, StreamExt};
+    use futures::StreamExt;
     use hydro_deploy::Deployment;
     use serde::{Deserialize, Serialize};
     use stageleft::q;
@@ -2715,97 +2553,5 @@ mod tests {
         deployment.start().await.unwrap();
 
         assert_eq!(external_out.next().await.unwrap(), 1);
-    }
-
-    #[tokio::test]
-    async fn reduce_keyed_watermark_filter() {
-        let mut deployment = Deployment::new();
-
-        let flow = FlowBuilder::new();
-        let node = flow.process::<()>();
-        let external = flow.external::<()>();
-
-        let node_tick = node.tick();
-        let watermark = node_tick.singleton(q!(1));
-
-        let sum = unsafe {
-            node.source_iter(q!([(0, 100), (1, 101), (2, 102), (2, 102)]))
-                .tick_batch(&node_tick)
-                .reduce_keyed_watermark(
-                    watermark,
-                    q!(|acc, v| {
-                        *acc += v;
-                    }),
-                )
-        }
-        .all_ticks()
-        .send_bincode_external(&external);
-
-        let nodes = flow
-            .with_process(&node, deployment.Localhost())
-            .with_external(&external, deployment.Localhost())
-            .deploy(&mut deployment);
-
-        deployment.deploy().await.unwrap();
-
-        let mut out = nodes.connect_source_bincode(sum).await;
-
-        deployment.start().await.unwrap();
-
-        assert_eq!(out.next().await.unwrap(), (2, 204));
-    }
-
-    #[tokio::test]
-    async fn reduce_keyed_watermark_garbage_collect() {
-        let mut deployment = Deployment::new();
-
-        let flow = FlowBuilder::new();
-        let node = flow.process::<()>();
-        let external = flow.external::<()>();
-        let (tick_send, tick_trigger) = node.source_external_bincode(&external);
-
-        let node_tick = node.tick();
-        let (watermark_complete_cycle, watermark) =
-            node_tick.cycle_with_initial(node_tick.singleton(q!(1)));
-        let next_watermark = watermark.clone().map(q!(|v| v + 1));
-        watermark_complete_cycle.complete_next_tick(next_watermark);
-
-        let sum = unsafe {
-            let tick_triggered_input = node
-                .source_iter(q!([(3, 103)]))
-                .tick_batch(&node_tick)
-                .continue_if(tick_trigger.clone().tick_batch(&node_tick).first());
-            node.source_iter(q!([(0, 100), (1, 101), (2, 102), (2, 102)]))
-                .tick_batch(&node_tick)
-                .chain(tick_triggered_input)
-                .persist()
-                .reduce_keyed_watermark(
-                    watermark,
-                    q!(|acc, v| {
-                        *acc += v;
-                    }),
-                )
-        }
-        .all_ticks()
-        .send_bincode_external(&external);
-
-        let nodes = flow
-            .with_default_optimize()
-            .with_process(&node, deployment.Localhost())
-            .with_external(&external, deployment.Localhost())
-            .deploy(&mut deployment);
-
-        deployment.deploy().await.unwrap();
-
-        let mut tick_send = nodes.connect_sink_bincode(tick_send).await;
-        let mut out_recv = nodes.connect_source_bincode(sum).await;
-
-        deployment.start().await.unwrap();
-
-        assert_eq!(out_recv.next().await.unwrap(), (2, 204));
-
-        tick_send.send(()).await.unwrap();
-
-        assert_eq!(out_recv.next().await.unwrap(), (3, 103));
     }
 }
